refactor: inline_checkbox/inline_choice を prompt_util.py から picker.py に切り出し
