+++ src/redi/cli/config_command.py
@@ -1,7 +1,7 @@
 import argparse
 
 from redi.cli.alias import resolve_alias
-from redi.cli.prompt_util import inline_choice
+from redi.cli.picker import inline_choice
 from redi.config import (
     SUPPORTED_LANGUAGES,
     create_profile,
